[iOS] Updates NTP constraints for iPad
This CL updates the constraints of the fake omnibox displayed on the NTP
when there is no multitasking.
When the NTP is entering multitasking, the constraints can be changed.
When leaving multitasking they aren't updated. This CL fixes it.

Bug: 914538
Change-Id: I53b9e812c35c9f45c69d1deb459485c17b6d3903
Reviewed-on: https://chromium-review.googlesource.com/c/1382492
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#617473}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b559ade6647d2fc478284a0fb1209157ab4d40c3