[iOS] Switch |didFinishNavigation| to operate on NavigationContext.
This change prevents an edge case where |_navigationStates| contains a
non-null NavigationContext for the key |nil|, which can happen in
same-document back/forward navigation. This NavigationContext is
incorrectly used later when calling [self didFinishNavigation:nil],
when the caller expects there to be no navigation context.

Outside of this pathological case, this refactor is a no-op as it
simply extracted the lookup of NavigationContext from WKNavigation*
from this method to the caller.

Change-Id: If900a7cbed92d68a415ab2be8575f3e866f742ca
Bug: 917162
Reviewed-on: https://chromium-review.googlesource.com/c/1387951
Commit-Queue: Danyao Wang <danyao@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#618589}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8f2187bd41d8aa8ef5b2027d6beffcb25f48f7ea