[ios] Use ScopedObserver for infobar in SnapshotTabHelper
ScopedObserver is the preferred way of observing. It has the benefit of
automatically removing the object from observing all sources when that
object is destroyed.

This CL removes an unnecessary private class used for observing and
replaces it with ScopedObserver in SnapshotTabHelper.

Change-Id: I06d30f645ae97a3ea44bf83664755711c2777efb
Reviewed-on: https://chromium-review.googlesource.com/c/1390099
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#618943}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 592396234dbf2945b78a5abc8c2c7c494853a703