[iOS] Removing command ids in Google services settings
The mediator implements GoogleServicesSettingsServiceDelegate instead
of the coordinator.
The mediator receives the cell item when the user taps on a cell or
toggle a switch. The command is now dispatched, by the mediator
according to the item type.

Change-Id: I326c65110d6e5a1ecd8d1ac9ab337b6b8e9d6e21
Reviewed-on: https://chromium-review.googlesource.com/c/1390010
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#619158}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1fc1b4e5217db51ede96846146c986901c5ae3b0