[ios] Remove use of default snapshot image
The default snapshot image is a simply a 2x2 white pixel image. This image is
unnecessarily passed around and unnecessarily complicates code.
This CL removes use of a default snapshot image, and updates the API
contract to use nil as the sentinel value for failed retrieval or
generation.

Bug: 918032
Change-Id: Ie7d8312cad6d05912474c6205c1bbb4ec9d242e0
Reviewed-on: https://chromium-review.googlesource.com/c/1392264
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#619565}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe0e620bfc70cbadc9fb99509c4b945a287fe417