[iOS] Delete unused method BuildFakeGaiaCookieManagerServiceWithOptions.
iOS tests don't use BuildFakeGaiaCookieManagerServiceWithOptions, so it
is dead code and can be deleted.

This functionality can always be added later if/when there is a test
which needs it.

This is part of step 1a of the plan to eliminate FakeGCMS entirely:
https://docs.google.com/document/d/1t0ZtuV7h-znzdItFgBW0aKPscAwWXIBuNZnNlEGgi7g/edit

Bug: 907782
Change-Id: Ia924bd8f41e6f7090646eae02841a01743f104a2
Reviewed-on: https://chromium-review.googlesource.com/c/1384370
Commit-Queue: Lowell Manners <lowell@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#619644}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 357909fe63f449e025b97561d3143bb19d1d2b7b