[ios] Adds helper classes for observing WebStates in a WebStateList.
AllWebStateObservationForwarder forwards observer methods for all the
WebStates in a WebStateList, keeping track of WebStates as they are
added, removed, and replaced.

ActiveWebStateObservationForwarder forwards observer methods for just
the currently-active WebState, even as the active WebState in the
WebStateList changes.

These classes are intended to remove some of the boilerplate bookkeeping
that is necessary when trying to perform either of these common tasks.

Change-Id: I8320f1d1c6aef180a79adf9cbe614cd2066b84fb
Reviewed-on: https://chromium-review.googlesource.com/c/1393531
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#619650}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cc11e7bffa4d92221f92973d9b61645a800befcd