Remove -[TabModel navigationCommittedInTab:previousItem:].
The body of the method was simplified and merged into the caller. Most
of the old code was related to calculation of "is in-page navigation"
boolean flag, which is actually present in LoadCommittedDetails.

New code simply uses LoadCommittedDetails.is_in_page flag.

The motivation for the change is to eventually replace didCommitNavigationWithDetails
with didFinishNavigation. didFinishNavigation does not provide
"previous navigation item" pointer, which was used in the old code.

Bug: 661983
Change-Id: Ida8ffa4c1127a2750be642e29bb86e90d42dfad9
Reviewed-on: https://chromium-review.googlesource.com/c/1396451
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#620466}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b117c3202ed77ed13e450819e608b647da6fc4b1