Check the proper URL when deciding if the password manager available for HTTP proxy auth.
ChromePasswordManagerClient currently checks if the password manager available on the page.
However, the "current page" doesn't make sense as the password may be saved for any of the frames or proxy.
Specifically for proxy the current logic is broken because the committed page has nothing to do with the proxy address.
In fact every caller of ChromePasswordManagerClient::IsFillingEnabledForCurrentPage know what URL they want to save a password for. This URL is passed to the client.

Bug: 916165
Change-Id: I0a39dfebc74e14e06553d8cb657709244b87f3d7
Reviewed-on: https://chromium-review.googlesource.com/c/1398444
Commit-Queue: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: John Wu <jzw@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#621180}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 18dbbac0817f9d10a0b48ca3f2bceaa656c743ae