Replace SigninManager by IdentityManager in AccountConsistencyService
It was using the SigninManager API to observe primary account sign-in
status changes and also to verify whether a primary account was set or
not. The IdentityManager API is a good replacement for those use cases.

Note that the class is still using GaiaCookieManagerService but in
order to move that code away from that dependency the unit test needs
to be converted away from any usage of ProfileOAuth2TokenService. That
will be done in a follow-up patch.

Bug: 797931
Change-Id: I0820d0c229dbba3ec2c2df1c9432f6275571db45
Reviewed-on: https://chromium-review.googlesource.com/c/1394525
Commit-Queue: Sergio Villar <svillar@igalia.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#621536}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8853032ec0edd9f4656c7abaeda4c504934c6119