Fix downloadManager cookies usage.
Set cookies on the NSURLConfiguration before the session is created,
because once the session is created the NSURLConfiguration object can't
be modified (NSURLSession returns a copy of the configuration not the
one that is used by the session).

This is a resubmit of cr/1394924 with disabling the cookies test for
iOS 10, because the NSHTTPSharedCookieStorage has a different default
cookiesAcceptPolicy.

Bug: 916662
Change-Id: I6f6c298bf65def96775867ce7eb973de45afbd12
Reviewed-on: https://chromium-review.googlesource.com/c/1394924
Reviewed-on: https://chromium-review.googlesource.com/c/1413113
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#623297}
Cr-Original-Original-Commit-Position: refs/heads/master@{#621502}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: da9b2b5bc53094a54107bba9fe80e364df191657