[Nav Experiment] Add workaround for WKWebView history clobber bug.
The clobber bug happens when the error page (loaded via -loadHTMLString)
becomes associated with the WKBackForwardListItem after the original
item the error page was loaded for, after the original item has been
reloaded in online mode.

This workaround detects when clobber is about to happen and uses
restore session to rewrite the clobbered entry and the forward history
after it. This works well if the clobber is triggered by a forward
navigation, as in the repro steps in crbug.com/887497.

Also added UMA metrics to track the frequency of this bug in the
real world.

Bug: 887497
Change-Id: I2b69d0ed8f8b9ad984b7c29aa6cb61be887f57e5
Reviewed-on: https://chromium-review.googlesource.com/c/1407607
Commit-Queue: Danyao Wang <danyao@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#623397}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ef7c2b2c5189e104d0f62d0da7ffbc555a5e0106