[iOS] Ensure correct URL is shown while a SafeBrowsing warning is displayed
SafeBrowsing warnings are shown in between decidePolicyForNavigationAction
and didStartProvisionalNavigation. For renderer-initiated and history-initiated
navigations, we don't normally update the displayed URL by this point, so
the warning is displayed with the previous URL still in the omnibox, which
is confusing for users.

To avoid this confusion, we need to create a transient navigation item for
the SafeBrowsing warning. However, WKWebView embedders do not receive any
callback when the warning is displayed. So this CL uses polling to detect
the warning. Specifically, after decidePolicyForNavigationAction, we check
for the warning every 20ms, until we next receive a WKNavigationDelegate
callback or loading is cancelled. When the warning is detected, we create
a transient navigation item.

Bug: 911687
Change-Id: I1c9513858e0152bf28d2b1204b6a34ee3fb448c9
Reviewed-on: https://chromium-review.googlesource.com/c/1412732
Commit-Queue: Ali Juma <ajuma@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#623483}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f04c02d70a91f3522fe23c26ed8441ffed8dde6