Remove ChromeWebViewFactory
Chrome no longer uses SSOAuth with UIWebView (it now uses WKWebView to show the Gaia add
account page).

The sign-in code was the last component using the UIWebView. This means that ChromeWebViewFactory
is not obsolete and it should now be removed. This CL does this clean-up.

Bug: 922979
Change-Id: I85486a19f21d0e1ad6917cf544b0d10b0472e71b
Reviewed-on: https://chromium-review.googlesource.com/c/1420918
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#624560}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 48548608c2cdbd9668383907d8ba6f35e6d53655