Use SyncService instead of ProfileSyncService when possible
The sync service API is exposed by the syncer::SyncService interface
and browser_sync::ProfileSyncService is an implementation of it. The
code should prefer to use the abstract interface (for one, testing
is easier as the surface to mock is smaller) instead of a particular
concrete implementation.

This change both factories in src/ios to return a syncer::SyncService
and port client code that could not use syncer::SyncService to access
the service as browser_sync::ProfileSyncService via dedicated methods.

Those methods will eventually be removed once there are alternative
way to access the corresponding API (either because it is moved to a
separate service, or moved to the syncer::SyncService interface).

Bug: 924508
Change-Id: I49db34ea7c79ef7b6165eb6d51f03b0af0394b41
Reviewed-on: https://chromium-review.googlesource.com/c/1429939
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#625443}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 780f6874acdc3c5d6b7a752c58b094062a344977