[ios] Do not hide infobar in side swipe tab change
The goal of this CL is to fix side swiping so that the infobar
doesn't magically disappear when side swiping.

Previously, the infobar was being hidden when the user side swipes the
omnibox to change the tab. While it is necessary to hide the infobar
and other accessory UI when swiping to navigate, it is not necessary
for swipe to change tabs. The reason it is not necessary is that the
whole UI is covered by the side swipe horizontal carousel cards (which
use a snapshot of the tab). On the other hand, side swipe to navigate
does not place a snapshot above the tab, so the tab is sliding behind
a stationary infobar.

This CL removes unnecessary code that hides accessory views when
side swiping the omnibox.

Bug: 925246
Change-Id: I767566d0b4ac7e8371b1ce69368855d005433d64
Reviewed-on: https://chromium-review.googlesource.com/c/1435721
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#626377}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a438603903b8033fd324f1fc3161f8f98fde0c0a