[ios] Move SadTab placement using NamedGuide out of BVC
The goal of this CL is to reduce SadTabCoordinator code in BVC so that
SadTabCoordinator can eventually be moved to BrowserCoordinator.

NamedGuides decouple placement of views from BVC. This CL moves SadTab
view placement out of BVC and next to the presentation code so that
presentation and placement are adjacent to one another. This helps to
reduce the SadTabCoordinator footprint in BVC.

Bug: 905524
Change-Id: Ib6744e145380441b3b43c783a12162a9bc0f7f1b
Reviewed-on: https://chromium-review.googlesource.com/c/1426176
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#626450}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 479b677b3d38fbebf013a5cd4da7296ad51ab4cd