[iOS] Fixed omnibox state being corrupted after presenting js alert
This CL fixes a bug where the omnibox state would be incorrect after
presenting a javascript alert in the middle of opening a new tab.
The root cause is that the omnibox focus methods are called in the
middle of a previous animation, so they were ignored. To fix this, we
store the expected final state of the omnibox, and set the state to that
after the animations finish.

Bug: 921390
Change-Id: Ibd4e8e857621344e4397153400e6dc4b43617c5c
Reviewed-on: https://chromium-review.googlesource.com/c/1434285
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#626618}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: df5b958678c91506562ccea5088ebefe473fc785