[ios] Fix download bar disappearing on side swipe
The download bar was disappearing on side swipe. During side swipe, the
content view is removed from the view hierarchy and re-added when the
side swipe completes. This removal and re-adding is causing an adverse
side-effect in which the download bar's bottom constraint is not enforced.

This CL stops the removing and re-adding of the content area for the
following reasons:

1) The optimization of removing the content area from the view hierarchy
is not worth the complexity it adds to the code. Side swipe is not a
highly used feature.

2) Since the overlaid side swipe view fully covers the screen and is
fully opaque, UIKit will already optimize the drawing calculations.

3) The content area was merely removed from the hierarchy, not from
memory.

4) The content area is presented through view controller containment.
Simply removing the view without dismissing the view controller is
uncommon, and may cause issues.

Bug: 926611
Change-Id: Idab9bcd183d1471edbc4aff31324c9e34277bcb5
Reviewed-on: https://chromium-review.googlesource.com/c/1445234
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#627631}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c6d7461e53e5cffffe6fa6e8a1886bfe0acf2080