[ios] Move TopAlignedImageView to c/b/ui/elements
TopAlignedImageView was previously only used in TabGrid. It will now
also be used when adding a black/white placeholder when a tab is evicted.

This CL moves the files so that it can be imported by another directory
in the future.

Change-Id: Idb7692f95192befafc2d0cdb0966df800a10fddc
Reviewed-on: https://chromium-review.googlesource.com/c/1448520
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#628353}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ae0d7ac320145502e9f6102d48f0f0baedb8593d