[ios] Remove snapshot in prerender
There is no reason for taking a snapshot in prerender.
Here are the possible outcomes:

1) If the prerender tab is used, then the tab will be displayed.
A snapshot will be taken when the tab is switched, tab switcher is
displayed, omnibox swipe is used, or navigation has completed.
This is the same as for any normal tab.

2) If the prerender tab is canceled and never used, there will never
be any use for the taken snapshot.

3) There is nowhere that the prerender snapshot is used while the user
is typing in the omnibox.

This CL removes taking a snapshot in prerender. One marginal benefit
is that less memory and disk space is used to create prerendered
snapshots. These may have previously been left on disk until purge.

Bug: 917929
Change-Id: I8dde01f478e5a5134ca879fb86fd5b7be944b097
Reviewed-on: https://chromium-review.googlesource.com/c/1450975
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Auto-Submit: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#628629}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ff3805a845470eac86d6c62a66693c9cff2d18a9